+++ include/app/ModExp.h
@@ -5,7 +5,7 @@
 #ifndef MONEXP_BASIC_H
 #define MONEXP_BASIC_H
 
- int modExp(int a, int e, int n);
+ __int128 modExp(__int128 a, __int128 e, __int128 n);
 
 #endif
